test-requirements.txt
pytest
pytest-django
pylint
pylint-django
model_mommy
